Add in cognitive only and sample app class examples
