+++ Arc/Components/DefaultImpl/SampleAppNavigationController.swift
@@ -25,7 +25,6 @@
 //
 
 import Foundation
-import Arc
 import UIKit
 
 open class SampleAppNavigationController : AppNavigationController {
